Added die_cast.die()
Added an integer based die roller so scripts won't have to convert
die expressions to a string just to roll them. One can now just say a
variable = dc.die(6,2) for a simple 2d6 roll, rather than dc.cast("2d6").
This was more for script writing, since cast was designed as a method to
have a script that could take a string direct from individuals, but down the
line an interface might take care of all this and do die rolls itself with
die.